Moved the Checks to a different Class
