+++ UnitySDK/Assets/ML-Agents/Scripts/CoreBrainInternal.cs
@@ -41,7 +41,7 @@ namespace MLAgents
 #pragma warning restore
 
 
-        private List<string> tmp_editor_errors;
+        public List<string> currentFailedModelChecks;
         
         Batcher brainBatcher;
 
@@ -102,7 +102,6 @@ namespace MLAgents
         {
             if (model != null)
             {
-
                 Debug.Log("Initialize");
                 InferenceEngineConfig config;
                 m_engine = InferenceAPI.LoadModel(model, config);
@@ -111,13 +110,33 @@ namespace MLAgents
                 inferenceInputs = GetInputTensors();
                 inferenceOutputs = GetOutputTensors().ToArray();
 
-//                _modelVersionNumber = GetModelData(m_engine, _nodeNames.VersionNumber);
+                _modelVersionNumber = GetModelData(m_engine, _nodeNames.VersionNumber);
                 _memorySize = GetModelData(m_engine, _nodeNames.MemorySize);
-//                _isContinuous = GetModelData(m_engine, _nodeNames.IsContinuousControl);
+                _isContinuous = GetModelData(m_engine, _nodeNames.IsContinuousControl);
+
+                currentFailedModelChecks = new List<string>();
+                if (_modelVersionNumber != version)
+                {
+                    currentFailedModelChecks.Add("Incompatible Version");
+                }
+                if (_memorySize == -1)
+                {
+                    currentFailedModelChecks.Add("No Memory Size");
+                }
+                if (_isContinuous == -1)
+                {
+                    currentFailedModelChecks.Add("Could not infer action space type from model");
+                }
+                currentFailedModelChecks.AddRange(InternalBrainTensorCheck.GetChecks(
+                    m_engine, inferenceInputs, inferenceOutputs, brain.brainParameters,
+                    _isContinuous, _memorySize > 0, _nodeNames));
             }
             else
             {
                 m_engine = null;
+                currentFailedModelChecks = new List<string>();
+                currentFailedModelChecks.Add(
+                    "There is no model on this Brain, cannot run inference. (But can still train)");
                 // TODO : Implement
                 throw new UnityAgentsException("ERROR TO IMPLEMENT : There was no model");
             }
@@ -143,49 +162,9 @@ namespace MLAgents
             }
             catch
             {
-                Debug.LogError("Node not in graph " + name);
+                Debug.Log("Node not in graph " + name);
                 return -1;
             }
-        }
-
-        public List<string> GetModelChecks()
-        {
-            if (m_engine == null)
-            {
-                return new List<string>();
-            }
-            
-            // TODO : When the method is called, the engine and the brainParameters must be up to date
-            inferenceInputs = GetInputTensors();
-            inferenceOutputs = GetOutputTensors().ToArray();
-
-            var failedChecks = new List<string>();
-            if (_modelVersionNumber != version)
-            {
-                // TODO : Make better
-                failedChecks.Add("Model has not been trained using the same version as the " +
-                           "Internal Brain");
-            }
-
-            if (_isContinuous == 1 &&
-                brain.brainParameters.vectorActionSpaceType != SpaceType.continuous)
-            {
-                // TODO : Implement
-//                errors.Add("Error to implement");
-            }
-            failedChecks.AddRange(CheckInputTensorShape(
-                inferenceInputs,
-                brain.brainParameters,
-                _nodeNames));
-            failedChecks.AddRange(CheckInputTensorPresence(
-                inferenceInputs,
-                brain.brainParameters,
-                _nodeNames));
-            failedChecks.AddRange(CheckOutputTensorPresence(
-                inferenceOutputs,
-                brain.brainParameters,
-                _nodeNames));
-            return failedChecks;
         }
 
         private Tensor[] GetInputTensors()
@@ -240,7 +219,7 @@ namespace MLAgents
                       Name = _nodeNames.VectorObservationPlacholder,
                       Shape = new long[2]
                       {
-                          n_agents, bp.vectorObservationSize * bp.numStackedVectorObservations
+                          n_agents, bp.vectorObservationSize * bp.numStackedVectorObservations 
                       },
                       ValueType = Tensor.TensorType.FloatingPoint
                   });
@@ -330,167 +309,6 @@ namespace MLAgents
 
         }
 
-        private static List<string> CheckInputTensorPresence(
-            IEnumerable<Tensor> tensors,
-            BrainParameters brainParams,
-            NodeNames nodeNames)
-        {
-            var result = new List<string>();
-            var tensorsNames = tensors.Select(x => x.Name);
-
-            // If there is no Vector Observation Input but the Brain Parameters expect one.
-            if ((brainParams.vectorObservationSize != 0) &&
-                (!tensorsNames.Contains(nodeNames.VectorObservationPlacholder)))
-            {
-                result.Add("The model does not contain a Vector Observation Placeholder Input. " +
-                           "You must set the Vector Observation Space Size to 0.");
-            }
-
-            for (var visObsIndex = 0;
-                visObsIndex < brainParams.cameraResolutions.Length;
-                visObsIndex++)
-            {
-                if (!tensorsNames.Contains(
-                    nodeNames.VisualObservationPlaceholderPrefix + visObsIndex))
-                {
-                    result.Add("The model does not contain a Visual Observation Placeholder" +
-                               " Input for visual observation "+visObsIndex+".");
-                }
-            }
-            return result;
-        }
-        
-        private static List<string> CheckOutputTensorPresence(
-            IEnumerable<Tensor> tensors,
-            BrainParameters brainParams,
-            NodeNames nodeNames)
-        {
-            var result = new List<string>();
-            var tensorsNames = tensors.Select(x => x.Name);
-
-            // If there is no Action Output.
-            if (!tensorsNames.Contains(nodeNames.ActionOutput))
-            {
-                result.Add("The model does not contain an Action Output Node.");
-            }
-
-            return result;
-        }
-
-        private static List<string> CheckInputTensorShape(
-            IEnumerable<Tensor> tensors, 
-            BrainParameters brainParams,
-            NodeNames nodeNames)
-        {
-            var result = new List<string>();
- 
-            var tensorTester =
-                new Dictionary<string, Func<Tensor, BrainParameters, string>>()
-                {
-                    {nodeNames.VectorObservationPlacholder, CheckVectorObsShape},
-                    {nodeNames.PreviousActionPlaceholder, CheckPreviousActionShape},
-                    {nodeNames.RandomNormalEpsilonPlaceholder, ((tensor, parameters) => null)},
-                    // TODO : Need a checker for some of those
-                    {nodeNames.ActionMaskPlaceholder, ((tensor, parameters) => null)},
-                    {nodeNames.SequenceLengthPlaceholder, ((tensor, parameters) => null)},
-                    {nodeNames.RecurrentInPlaceholder, ((tensor, parameters) => null)},
-                };
-
-            for (var visObsIndex = 0;
-                visObsIndex < brainParams.cameraResolutions.Length; 
-                visObsIndex++)
-            {
-                var index = visObsIndex;
-                tensorTester[nodeNames.VisualObservationPlaceholderPrefix + visObsIndex] =
-                    (tensor, bp) => CheckVisualObsShape(tensor, bp, index);
-            }
-            
-            foreach (var tensor in tensors)
-            {
-                if (!tensorTester.ContainsKey(tensor.Name))
-                {
-                    result.Add("No placeholder for input : " + tensor.Name);
-                }
-                else
-                {
-                    var tester = tensorTester[tensor.Name];
-                    var error = tester.Invoke(tensor, brainParams);
-                    if (error != null)
-                    {
-                        result.Add(error);
-                    }
-                }
-            }
-            return result;
-        }
-        
-        private static string CheckActionMaskShape(
-            Tensor tensor,
-            BrainParameters brainParams)
-        {
-            Debug.Log("error to implement.");
-            return null;
-        }
-
-        // TODO : Rename these check
-        private static string CheckVectorObsShape(
-            Tensor tensor,
-            BrainParameters brainParams)
-        {
-            var vecObsSizeBp = brainParams.vectorObservationSize;
-            var numStackedVector = brainParams.numStackedVectorObservations;
-            var totalVecObsSizeT = tensor.Shape[1];
-            if (vecObsSizeBp * numStackedVector != totalVecObsSizeT)
-            {
-                return string.Format(
-                    "Vector Observation Size of the model does not match. " +
-                    "Received {0} x {1} but was expecting {2}.",
-                    vecObsSizeBp, numStackedVector, totalVecObsSizeT);
-            }
-            return null;
-        }
-        
-        private static string CheckPreviousActionShape(
-            Tensor tensor,
-            BrainParameters brainParams)
-        {
-            var numberActionsBp = brainParams.vectorActionSize.Length;
-            var numberActionsT = tensor.Shape[1];
-            if  (numberActionsBp != numberActionsT)
-            {
-                return string.Format(
-                    "Action Size of the model does not match. " +
-                    "Received {0} but was expecting {2}.",
-                    numberActionsBp, numberActionsT);
-            }
-            return null;
-        }
-        
-        private static string CheckVisualObsShape(
-            Tensor tensor,
-            BrainParameters brainParams,
-            int visObsIndex)
-        {
-            
-            var resolutionBp = brainParams.cameraResolutions[visObsIndex];
-            var widthBp = resolutionBp.width;
-            var heightBp = resolutionBp.height;
-            var pixelBp = resolutionBp.blackAndWhite ? 1 : 3;
-            var widthT = tensor.Shape[1];
-            var heightT = tensor.Shape[2];
-            var pixelT = tensor.Shape[3];
-            if  ((widthBp != widthT) || (heightBp != heightT) || (pixelBp != pixelT))
-            {
-                return string.Format(
-                    "The visual Observation {0} of the model does not match. " +
-                    "Received Tensor of shape [?x{1}x{2}x{3}] but was expecting [?x{4}x{5}x{6}].",
-                    visObsIndex, widthBp, heightBp, pixelBp, widthT, heightT, pixelT);
-            }
-            return null;
-        }
-        
-        
-
         /// Uses the stored information to run the tensorflow graph and generate 
         /// the actions.
         public void DecideAction(Dictionary<Agent, AgentInfo> agentInfo)
@@ -559,14 +377,12 @@ namespace MLAgents
             if (EditorGUI.EndChangeCheck())
             {
                 InitializeModel(m_model);
-                tmp_editor_errors = GetModelChecks();
             }
             
             // TODO : Remove :
-            tmp_editor_errors = GetModelChecks();
 //            Debug.Log(GetModelErrors().Count);
             
-            foreach (var error in tmp_editor_errors)
+            foreach (var error in currentFailedModelChecks)
             {
                 if (error != null)
                     EditorGUILayout.HelpBox(error, MessageType.Error);
